fixing cost param in sale

+++ src/main/java/models/Sale.java
@@ -17,7 +17,6 @@ public class Sale {
     private int productId;
     private double cost;
     private int numPurchased;
-
     private DocumentReference documentReference;
 
     public static CollectionReference salesCollection;
@@ -39,7 +38,7 @@ public class Sale {
         this.productId = productId;
         int numPurchased = document.getLong(fields.numPurchased).intValue();
         this.numPurchased = numPurchased;
-        double cost = document.getLong(fields.cost).intValue();
+        double cost = document.getLong(fields.cost).doubleValue();
         this.cost = cost;
         this.documentReference = getSaleDocument();
     }
